+++ apps/chef_app/README.md
@@ -1,7 +1,11 @@
-# Devmesh Backend
+# Chef.ai Lite
 
-## Demo
+## Demos
 
+### Chef.ai Webapp with Flask Templates/Vanilla JS
+![Chef.ai in action!](./chef_ai_lite.gif)
+
+### Original [Chef.ai](https://github.com/fcr3/chef_devmesh) Webapp with React.js
 ![Chef.ai in action!](./chef_ai.gif)
 
 ## How to use the app:
@@ -12,91 +16,40 @@
 5. See recipes and choose from 10 max
 6. Repeat!
 
-## Important Details:
-- Backend Stack: Flask, Celery, Redis
-- The Flask backend is implemented asynchronously to accommodate for heavy traffic.
-- Use of task IDs and interval function calls are used to track status/result
-  1. Use initial AJAX call to get the task ID
-  2. At interval, use another AJAX call using the task ID to grab status
-  3. End interval checking once AJAX call responds with success/failure
-- The test react app demonstrates how to properly use the Flask backend
-- Object detection model based on images from the internet, labeled by Christian
-  - Mimics the general workflow of a ML software developer developing on Tensorflow
+## Setup
+All setup/startup is streamlined through the make commands!
+
+### Makefile
+- **make all** - does **make data** and **make deps**
+- **make data** - downloads test data
+- **make deps** - downloads the optimized OpenVINO models
+  - **make deps_FP16** and **make deps_FP32** are subroutines of **make deps**
+  - Commands download their respective models based on precision
+- **make run** - runs the flask backend and opens the webpage. Plug in your Intel NCS 2 in order to use `Detect (MYRIAD)`!
+- **make install-reqs** - downloads required packages
+- **make clean-lite** - deletes data/models
+- **make clean** - deletes data/models and required packages
+- **make help** - lists what the commands above do
+
+### How to set up Edamam Account to use API
+1. Go to Edamam API's website and register for an account. You will need to sign up for their recipe API as well as make an application that uses the recipe API.
+2. Copy the app ID and app key into the remaining key-value pairs, respectively into the beginning portion of the script section in `templates/index.html`. If you have trouble, do a text search for `CONFIGS` to find the config keys.
+
+## Details:
+- Backend Stack: Flask
+- Webpage is served via root route (127.0.0.1:5000)
+  - Requests to the backend are done via fetch HTTP requests
+- Synchronous Backend
+- You can find original repository [here](https://github.com/fcr3/chef_devmesh).
 
 ## Notable Dependencies:
 - Backend:
   - flask
-  - celery
-  - openCV
-  - OpenVINO 2019 R1.1
+  - PIL.Image
   - numpy
   - base64
-  - PIL.Image
-  - io
-  - redis
-- Frontend:
-  - React dependencies
-  - axios
-  - react-webcam
-
-## Setup
-
-### Installing Dependencies
-1. pip3 or pip install all the python dependencies for the backend
-2. Set up redis by following the installation in the backend setup section.
-2. `npm install` in the `test_react_app` folder to install all the dependencies for the react app.
-3. Download weights for food model and gesture model
-  - Food Model FP32: https://drive.google.com/open?id=1nDDaEpm7hNG8eR4vXBATMpWRrOL-4-jq
-  - Food Model FP16: https://drive.google.com/open?id=1eC-ZKpiSOW8pHZ1m3rIbnMfcKN2Hf1gI
-  - In the case that you don't like my models, please feel free to train your own models and plug them into the app. My recommendation is to follow the tutorial links below to train using Tensorflow's Object Detection API.
-
-### Procedure for starting application:
-1. open four terminals
-2. first terminal -> `$ redis-server`
-3. second terminal -> `$ cd <INSTALL_DIR>/devmesh_backend/ && celery worker -A openvino_backend.celery --loglevel=info`
-4. third terminal -> `python3 openvino_backend.celery`
-5. Make a config.js file and add the necessary requirements such as app_id, app_key, and Raspberry Pi backend url.
-5. fourth terminal -> `cd <INSTALL_DIR>/devmesh_backend/test_react_app && npm start`
-
-### Additional Details for config.js
-1. Make an account on dataplicity and go through their tutorial to setup your raspberry pi for access via the internet
-2. Use the special url and copy and paste that into the first key `BACKEND_URL`
-3. Go to Edamam API's website and register for an account. You will need to sign up for their recipe API as well as make an application that uses the recipe API.
-4. Copy the app ID and app key into the remaining key-value pairs, respectively
-
-### Common Errors:
-- Dealing with ELIFECYCLE Errors?
-  1. `sudo npm cache clean --force`
-  2. `sudo rm -rf node_modules`
-  3. `sudo npm install --save`
-  3. `npm start`
-- Dealing with modules not being found?
-  1. Pay attention to what you used to install dependencies. Pip and pip3 might not install to the python you might be expecting. In addition, using sudo rather --user may also lead to some referencing problems. Make sure to stay consistent!
-
-## References
-
-### Frontend References:
-- https://www.npmjs.com/package/react-webcam
-- https://www.npmjs.com/package/react-images-upload
-
-### Backend Setup:
-- https://blog.miguelgrinberg.com/post/using-celery-with-flask
-- https://redis.io/topics/quickstart
-
-### Object detection model tutorial:
-- https://tensorflow-object-detection-api-tutorial.readthedocs.io/en/latest/training.html
-- https://github.com/tensorflow/models/blob/master/research/object_detection/g3doc/detection_model_zoo.md
-- https://github.com/EdjeElectronics/TensorFlow-Object-Detection-API-Tutorial-Train-Multiple-Objects-Windows-10/issues/184
-
-## Past Research
-
-### OCR:
-- ???
-
-### PixelNet Detection:
-- https://github.com/banderlog/open_model_zoo/blob/24ca50034555721d876c8314ad36b4d6b03bf321/demos/python_demos/text_detection_demo.py
-
-### EAST Text Detection:
-- https://www.pyimagesearch.com/2018/08/20/opencv-text-detection-east-text-detector/
-- https://medium.com/@tomhoag/opencv-text-detection-548950e3494c
-- https://bitbucket.org/tomhoag/opencv-text-detection/src/master/
+  - openCV
+  - OpenVINO 2019 R2
+- Fronend:
+  - Vanilla Js (whatever Javascript comes with the browser)
+  - Regular HTML/CSS
